--- web/interceptor/helpers.go
@@ -1,32 +1,0 @@
-package interceptor
-
-import (
-	"fmt"
-)
-
-func has(method string) bool {
-	_, ok := accessRolesFor[method]
-	return ok
-}
-
-func CheckAccessMethods(expectedMethodNames map[string]bool) error {
-	missingMethods := []string{}
-	superfluousMethods := []string{}
-	for method := range expectedMethodNames {
-		if !has(method) {
-			missingMethods = append(missingMethods, method)
-		}
-	}
-	for method := range accessRolesFor {
-		if !expectedMethodNames[method] {
-			superfluousMethods = append(superfluousMethods, method)
-		}
-	}
-	if len(missingMethods) > 0 {
-		return fmt.Errorf("missing required method(s) in access control table: %v", missingMethods)
-	}
-	if len(superfluousMethods) > 0 {
-		return fmt.Errorf("superfluous method(s) in access control table: %v", superfluousMethods)
-	}
-	return nil
-}
